Move LandingPage folder into components folder
- I had mistakenly put LandingPage directly in the src folder, so I have
just moved it into the Components folder.
